Cleanup TODO in glance/gateway.py for elasticsearch being unavailable
In commit f53584c2688244f4f5b706aadac1875109a463c5 we added a TODO to be
more obvious about the error when elasticsearch isn't available if you
try to start the search service (since elasticsearch is an optional
dependency).  We couldn't add new errors at the time because of
StringFreeze.

This resolves the TODO by adding the new SearchNotAvailable exception in
place of the generic ServiceUnavailable exception.

Closes-Bug: #1441764

Change-Id: Ia300d9647f372a677dc536e750dd2ff064601c29