tests: don't hardcode strace usage
The wait_for_servers helper in glance.tests.functional.FunctionTest
uses the strace(1). While strace(1) is available on FreeBSD, but with a
huge limitation that it works only on i386 hosts.

Instead of always using strace(1), use it only on Linux, for FreeBSD use
the truss(1) tool, that is a native tracing tool that is a part of base
system, and in all other cases just don't try to obtain a trace.

Closes-bug: #1467110
Change-Id: I51cc43fe4ce71eed80bbf4f18f5b060b09494458