Adds os_admin_user to common OS image prop metadef
This will add os_admin_user to the list of common properties for
operating systems metadata definition.

Commit I215bb82668df074141397909b58fa9189f46b4da is adding
os_admin_user to the list of image metadata it understands.  As
a result, it's beneficial to also have that property defined
in the metadef.

Change-Id: I7cf70acf3c6a4f801b34d52926c444d05b79fd51
Implements: blueprint libvirt-set-admin-password