Fix duplicate unique constraint in sqlite migration
The sqlite migration 37 created duplicate unique constraints on
the image_properties table, one unnamed and one called
ix_image_properties_image_id_name.

This causes the unit tests to fail with sqlalchemy 1.0+ since it
catches the problem when it compares the model to the schema.

The fix is to only create the named constraint.

Change-Id: I7ee19f08f2c107987e763466dc6e3b07185598c9
Closes-Bug: #1467008