Remove usage of assert_called_once in mocks
This was never a real part of the mock api, but mock was happily mocking
it out and letting the tests pass. The new mock release made this
behaviour break, causing tests to fail on all patches. This patch
replaces instances of assert_called_once with an assert of the mock's
call_count.

Closes-Bug: #1473369
Co-Authored-By: Ian Cordasco <graffatcolmingov@gmail.com>
Change-Id: I12ff85a9480d8000b0de54939c3be71458162a40