V1: Stop id changes after the image creation
Currently in v1 if someone tries to update the ID after the image is
created it will cause a 500 to be returned (or if updating to the same
ID a 200). This now returns a 403.

ApiImpact

Change-Id: Ibb40632f027b2b84f88aaf0a971a0f18de350187
Closes-Bug: 1483353
Closes-Bug: 1483688
(cherry picked from commit be0a27082b53c85f47d3c82ef852f285d975f9b8)