Fixed version unequality artifact filtering
When filtered by the value of version field artifact list was ignoring
the filtering operators other then equality. I.e. it was possible to
list only artifacts having the version equal to some specific value, but
not the artifacts having the version less then or greater then some
value.

To implement the proper filtering a CompositeProperty comparator was
added which filters the ORM entities in the lexigraphical order of
version fields: first by version prefix, and then - if prefixes are
equal - by version suffix.

Closes-bug: #1480393
FastTrack

Change-Id: I67ae323e2b74745d794442d71b020becc494646b