Changes behaviour when an image fails uploading
Currently when an image fails being uploaded the task is left in the
failed state (with the internal sever error message) and the image is
left in the saving state.

This patch calls for the image to be deleted as the image is never seen
by the user, also returns a more specific task error message.

APIImpact

Change-Id: I15fbe9f0041407892144c0d26eadc4ed939e5abf
Closes-Bug: 1487026