Fixed an HTTP 500 on artifact blob upload
If the HTTP PUT request to upload a binary file to an artifact does not
contain a Content-size header, the length of the binary file should be
detected during the upload.

However, the unset content size variable was leading to an exception as
its value (None) participates in some math operations during the upload.
Setting the value to 0 (same is v2 API) fixes the issue.

Closes-Bug: #1488037
FastTrack

Change-Id: Iadbd97899445b58c2c5bad961bca61cd6ab60d1a