Fixes the possibility of leaving orphaned data
When Glance scrubber encounters an error while scrubbing image data
from the backend, it puts the image in deleted status while leaving
the image locations in pending_delete status with data still in the
backend. This results in dark data.

This change attempts to fix this by leaving the image in
pending_delete status if the scrubber fails to delete the data from
the backend. This presents the scrubber with an opportunity to
retry the scrubbing in the next run.

Change-Id: Ia6fa0008317cc0d619e5f212c0f043e078fbea55
Closes-bug: #1444767