Fixed non-owner write-access to artifacts
There was no check for write-access priviledges during all the
mutating operations with artifacts (updates, patches, deletes,
actiovations etc), so any user who has an access to an artifact could
modify it. Because of this, non-owners could modify public artifacts
which was a major security issue.

Now this is addressed and an appropriate set of tests added to
prevent possible regressions.

FastTrack
Change-Id: I2f4c2d70b74ae39ababed2ba9b97559ef5ea6d6c
Closes-bug: #1489902