Use oslo utils to encode exception messages
oslo.utils supports a method that convers exception messages to
unicode.
Glance previously had the same method but now we can use
oslo implementation in glance code and we don't need to
bother about supporting it in glance.
It also seems that utility function exception_to_str
in glance.common.utils was accidentally deleted but we didn't fix
it for code that use this utility.

Change-Id: Ib46a931da23492cde8c40869b2a1d718a7f0115b