Add 'deactivated' status to image schema.
New 'deactivated' status was introduced in Kilo release,
but it doesn't listed in available image statuses in the schema.

It leads to issues  on the client side, when it can't validate
the image with this status against the schema and returns the error.

Change-Id: I5ec264614ae7ecf54b846ad0600442a18c61d24c
Closes-bug: #1505218
Related-bug: #1505134
(cherry picked from commit 135a946a2d1a74dda67fcc35ab31c9d83d4d0c40)