Add the Docker container format
A Docker image is a tar archive typically containing a container
filesystem. In order to use the nova-docker compute driver nova and boot
docker instances glance support of the docker container format is
required.

This patch adds the Docker container format to the Glance configuration.

DocImpact: Docs will need to be updated to indicate that the docker
container_format is now one of the default container formats supported
by Glance.

UpgradeImpact: Adds 'docker' to the default list of container_formats.
Operators will no longer need to configure specifically to use docker
containers.

Closes-Bug: #1519402
Change-Id: Ifa206686086232a5599e7bc27a852bc5f3186ebc