Catch UnsupportedAlgorithm exceptions
For signature verification, only SHA-2 family hash algorithms are
used.  Some older platforms, with an older version of openssl (older
than 1.0.1) do not support SHA-2 hashes.

This patch skips the signature verification unit tests that require
SHA-2 support on systems that do not have SHA-2 support.  It also
catches the UnsupportedAlgorithm exception that is generated when a
system that does not have SHA-2 support tries to verify the signature,
providing better feedback to the user.

Change-Id: I11b2366e96d49a5d502d6fccb0b6f620eb6c381d
Closes-Bug: 1522525