Fix _wait_on_task_execution()
Attempt to fix a race condition in test_all_task_api of
glance.tests.integration.v2.test_tasks_api.TestTasksApi.
_wait_on_task_execution() must use eventlet.sleep() instead of time.sleep() to
give control to the pending server task, instead of blocking the whole process.

Note: The time module is not monkey-patched, so time.sleep() really hangs the
current thread for the specified duration. For an unknown reason, the test pass
in most cases, but always fail with testtools.

Change-Id: I785a7cf0d556ad72c443946adac3b4f5f361edd8
Closes-Bug: #1541487