Always use constraints
This flips the tox.ini to always use constraints installation for all
targets. It drops the extra -constraints targets in the process. This
makes it so there is no developer change required to operate in our new
world order.

Note that coverage and release-notes jobs cannot use constraints yet to
limitations of the tools we use. So, they continue to use the normal
non-constraint install.

Change-Id: I218f16c585b7b8a9ea73c6cf1c5b0aef2dcde015
Depends-On: Ia08c15914d9dbfc3f5dc04ca69a7be26efb6fe64