Creating or updating a image member in a list causes 500
This change catches a type error that can be raised if a user mistakenly
uses a list over a dict. This occurs in both adding a new member and
updating a member that has already been added.

This will cause a HTTPBadRequest to be raised with instructions how to
fix the problem.

Change-Id: I6af3e0ae45ee535859c4ad278ccf995643225585
Closes-Bug: 1551689