revert warnerrors before gate breaks
warnerrors used to show the errors in output when the docs generate
but now it looks like when this patch lands [0] it will cause the gate
to fail. I will push up a patch to fix all the docs problems but we
should revert this so we don't have a day when the gate is broken.

[0] https://review.openstack.org/#/c/229951

Change-Id: Ica33f1fd5acdc17a115d7a3bd81f8f4d79c1a809