Remove state transition from active to queued
After commit Ieb03aaba887492819f9c58aa67f7acfcea81720e it is no
longer necessary to maintain a state transition from ACTIVE to
QUEUED. The only time this happened was when someone removed
the last location from an image. Since this was removed we
don't need it anymore.

Also, updates the image state transition images and docs to
represent the current state transitions.

Closes-Bug: #1555448

Change-Id: I390ae2a7110f06046f38914e946139b3b729bd2d