Deprecate `show_multiple_locations` option
There are currently several ways to disable/enable multiple_locations.
One is through a global config option (show_multiple_locations) and the
other one is through a more granular RBAC using Glance's policies.

Maintaining two different ways to configure, enable and/or disable a
feature is painful for developers and operators. Given that we have a
more granular way to manage this feature and that it's the preferred one
given the provided ability to chose what each role can/cannot do, this
patch proposes to remove the global config option.

There have been concerns about this proposal mostly on the fact that
it'll remove the ability to "turn off" the entire feature with a single
option. This doesn't seem to be a strong enough motivation to warrant
the effort of maintaining these 2 options.

This patch marks the aforementioned option as deprecated.

UpgradeImpact
DocImpact

Co-Authored-By: Flavio Percoco <flaper87@gmail.com>
Co-Authored-By: Nikhil Komawar <nik.komawar@gmail.com>

Lite-spec: https://review.openstack.org/360220
Change-Id: I1c5cb7834c2cf3295e10bad7fd07cfacb8e4ac50