Clarify language used in glanceapi future section
As per the comments and agreement on
Iede37e61784b6a8a974ce0ae29ccfab510d8cc53 , this patch proposes the
concerned changes. The former patch was merged as the consensus on the
idea of the Images API future section was converged. This is merely a
simplification to the grammatical tone on the respective sections.

Change-Id: Iefbdeadb78b2a7b866495d4a8f3e7a84cee46fe2