WADL to RST migration (part 2 - images)
This brings the Images API Reference to the glance tree as part of the
effort described here:
https://wiki.openstack.org/wiki/Documentation/Migrate#API_Reference_Plan

This patch contains the content of the Images v2 API reference converted
to RST. The content has been corrected where I noticed divergences with
reality (which actually happened a lot, so this patch contains a lot of
revisions).

Tasks and metadefs are submitted as separate patches.

Change-Id: If69c469da7f28f090a2286ec0b115d9bcaac4fbf
Closes-Bug: #1550457
Closes-Bug: #1589264
Closes-Bug: #1592715