Version negotiation api middleware to include v2.3
Glance API is at v2.3 listed to be CURRENT version but the version
negotiation middleware lists only v2, v2.0, v2.1, v2.2.

Also, adds test class to versions tests.

The current versions tests contain a class to test the versions
response, and a class to test version negotiation, but this put us
in a situation where the 'versions' response was correctly updated
for v2.3, but the negotiation to map v2.3 -> v2 was not, thereby
making calls to /v2.3/whatever fail even though all tests in both
classes passed.

This patch adds a new test class that inherits from both of the
current test classes.  It contains tests to verify that all the
versions mentioned in the versions response are mapped correctly
so this kind of partial update situation won't happen again.

Closes-Bug: 1609571

Co-Authored-By: Nikhil Komawar <nik.komawar@gmail.com>
Co-Authored-By: Brian Rosmaita <brian.rosmaita@rackspace.com>

Change-Id: I4018af408fa45f3ac0ad6e9c8229428a9f87089f
(Cherry-picked from commit 0d1daf44efed10526f582c16ebab56ffa53bc1da and
commit 8c3560bbf6fe9f5267d7cf3f087201540a1ebff0)