Add guideline to refresh config files shipped with source
This commit adds documentation to help authors identify the steps they
need to take while updating the config files and thereafter.

This documentation is being added as per question asked on change
I534a0dc2d9f82ab70517cde12bfd40d4f5ae69e2

There is some discussion as to when we will be getting rid of the config
file from the tree on the change
Iee695870aefc9cb6142dbc7400672b335588fcc9

Until we do get rid of the config files from source, we need to document
this process for any further updates to configs.

Change-Id: I2e3d43e43e00cd2a13b5a0fb794ac6e6d2e24dc2