Remove "Services which consume this" section
In the first few improvements to the help text of configuration
options, the section "Services which consume this" was included
to clearly indicate which services consume a particular
configuration option. However, this is redundant information as
all glance services have their configuration files. And, the very
fact that a configuration option appears in a certain file
indicates that the corresponding service potentially consumes it.

Hence, in this patch we drop the redundant section from the help text
of configuration options it appears in. The configuration options
that are impacted are:
* public_endpoint (glance/api/glare/versions.py)
* owner_is_tenant (glance/api/middleware/context.py)
* admin_role (glance/api/middleware/context.py)
* allow_anonymous_access (glance/api/middleware/context.py)
* max_request_id_length (glance/api/middleware/context.py)
* public_endpoint (glance/api/versions.py)
* image_cache_driver (glance/image_cache/__init__.py)
* image_cache_max_size (glance/image_cache/__init__.py)
* image_cache_stall_time (glance/image_cache/__init__.py)
* image_cache_dir (glance/image_cache/__init__.py)
* image_cache_sqlite_db (glance/image_cache/drivers/sqlite.py)
* admin_role (glance/scrubber.py)

Change-Id: I9ea635368994a9f89bb4f19a82104499e5174b46
Partial-Bug: #1570946