Allow member creation when using db.simple api
Within simple/db/api.py, DATA['members'] was being statically
initialized as a dict.

Inspection of the code revealed a reset function that initialized the
same as a list. Further code inspection seemed to confirm that it was
always intended to be a list.

Updated static initialization to be a list.

Change-Id: Ia53890bcd1161b66a1b365bb3bcc7fc9fbcf6e40
Closes-Bug: #1620833