Add note to docs on release notes prelude section
This commit adds a note to the developer documentation for adding a note
to the prelude section.

It was mentioned in this change [1] that adding notes to the prelude
section for those that are not significantly important isn't
necessary. So, this commit adds a general notice so the docs about the
same.

[1] https://review.openstack.org/#/c/364586/

Co-Authored-By: Nikhil Komawar <nik.komawar@gmail.com>
Co-Authored-By: Brian Rosmaita <brian.rosmaita@rackspace.com>

Change-Id: I0c507706ec39b5eafcdebfb3e1f57dc56976d9a5