Add node_staging_uri and enable_image_import opts
The node_staging_uri will be used for the API logic saving the data
user uploads to the new staging endpoint and async flow to pick it
from there when processing the rest of the workflow.

For the first iteration, only FS path is allowed and supported.

URI format used to provide future expansion for possibility to use
different shared backends between the nodes. The location access must
be shared between the nodes.

The enable_image_import config option is introduced as deprecated and
is only there to ensure smooth upgrades. Having it disabled per default
in Pike allows glance-api to start and operate with Newton config
files and the new feature disabled.

From Queens onwards 'node_staging_uri' option must be configured for full
operation of glance-api as the enable_image_import will default to True.
At this point the 'enable_image_import' option will become redundant
and can be removed on R.

Change-Id: Ie5ab21ec3f9c880fa042dcc68865c1fceb9463ec