Do not use service catalog for cache client
Previously cache client used endpoints from service catalog to
request cache items from node. It works perfectly on devstack
but it doesn't work when an OpenStack cloud is deployed with HA
(3/5/7 controllers). We need to use -H option instead.
This option makes cache management client interact with the local
endpoint which is more appropriate as image cache is local to
each node/controller.

Change-Id: I355fb36d9af7e06dcac70b8263dc1a2b915ffc8f
Closes-Bug: #1634486