Allow specifying OS_TEST_PATH (to reduce tests ran)
All the other main openstack projects support this and
I am trying to provide it for usage in our jenkins so that
our jenkins only runs the unit tests.

Setting this up in jenkins hit a snag though because it
appears that glance doesn't support OS_TEST_PATH (this
change allows it to).

Change-Id: I602c78dc6ef882a00592261aaf39a77215c31226