Skipping tests for location 'add', 'replace' on 'queued' images
Two of test cases added for the changes [1] and [2]
have been failing, throwing a webob.exc.HTTPBadRequest due to the
fake store url "http://storeurl.com/container/fake_location_1"
being used by both the failing test cases.
The faking in these tests seem to override the usage of
FakseStoreApi and instead call glance_store.

Since, the 2 tests that are failing are caused due to issues
with the faking and not due to the changes [1] and [2] itself,
for which they were added, this patch proposes to skip the
2 failing tests until they are resolved inorder to prevent
other changes from getting blocked.

[1] https://review.openstack.org/#/c/366995/
[2] https://review.openstack.org/#/c/324012/

Refrence Bug: #1649300

Change-Id: Ib70ce08528ef92d5fb488e4deb5fe82a7f74a634