refactor glare plugin loader tests to not mock private methods of stevedore
This change updates the glare plugin loader tests to avoid mocking
private methods of objects coming from the stevedore library. It looks
like some of the tests could probably be deleted, since they seem to be
testing stevedore and not glance, but it was easier to just fix them
than to verify that.

Change-Id: I5ee634d0af9d7942bfd5b69c3e7336876ab42269
Signed-off-by: Doug Hellmann <doug@doughellmann.com>