Correct 2.5 minor version bump release note
Current release note says adding a member to a private image will
result in a 409, but it actually results in a 403.  Changing the text
to say the more neutral "4xx response" without specifying a return
code, because there are several errors that could happen here, for
example, 401, 404, 403.  Interested parties will know to look in the
api-ref, where the error codes can be listed with more context than
is appropriate for a release note.

Change-Id: If513e5a0441b2c110c9339544a47ad33d77afaaf