Fix scrubber test failing py35 gate.
Recently added `test_scrubber_app_queue_errors_not_daemon` test
in glance/tests/functional/test_scrubber.py fails the non-voting
`gate-glance-python35-db-nv` gate job because in python3.5, the
output of subprocess call is a bytes object instead of string.
In preparation for enabling voting on this gate, fixed the failure
by converting the `err` object that is being checked to a string.

Change-Id: If84d572b192758da9dec743678f9f0cb0d563faa