Cleanup 'ResourceWarning: unclosed file' in py35.
Glance tests use a 'Server' class defined in 'glance.tests.functional.
__init__.py'. This object has a 'dump_log' method used by all derived
objects to write out the contents of a log file to the /tmp area.
The method does not close the file pointer after opening it for reading.
This causes a multitude of warnings of the type specified in title above
to appear when running under python3.5. This patch adds closing of the
file after the log dump completes.

Change-Id: I95703fef5d73716c3a09cba6d220bfd7519ad713