Restore man pages source files
Partially reverts I6d67e7a40e16468bb7bf4ac742361fb44eec4e28 to
restore the source files for the man pages.  These are used by
some packagers (confirmed that this is the case for Ubuntu) to
generate the man pages included in their distributions.

Does not add the man pages back to the HTML index, so these will
be orphans for now.  A follow up patch will either add a man build
to tox, or add the HTML generation back so we can review generated
content.

Long term fix will be to merge the content of the --help text
in the cmd files with the content of the man pages and figure
out how to generate the man pages from the --help text.  Right
now, the --help text is better in some regards, and the man pages
are better in others.

Change-Id: I2aa85b63a196c75224630981b322fecbf36f1b7c