Clean up acceptable values for 'store_type_preference'
The store values 'filesystem' and 'vmware_datastore' were
deprecated for 'store_type_preference' as per [1] in Newton.
This was to maintain uniformity with the values accepted for
'store' opt and 'store_type_preference' opt.

This patch removes acceptance for the old store values
and ensures uniformity between store names used in glance
and glance_store.

[1] I3d8593cdae6780ae721afedcf47dd45afa684f25

Change-Id: Id225275f7aad308bde450bef8bc6fd6ad7e3c6d2