Fix glance image-download error
When use command 'glance image-download --file xxx.raw <image-id>'
to download the image, it will fail with error code 500. It shows
"TypeError:'ImageTarget' object does not support item assignment"
in glance-api's log.

It is because that when instance snapshot of volume-backed instance
is uploaded to glance, image size is set to 0 from nova. While
downloading that snapshot with using image-download call it tries
to get the image size from the cache where it fails as image is
instance of policy.ImageTarget. Made change to set image size
on image_meta.target if image is a policy.ImageTarget.

Change-Id: I1a08baa465d3927b721678a5a7ab5f54304993d6
Co-Authored-By: Abhishek Kekane <abhishek.kekane@nttdata.com>
Co-Authored-By: Eric Fried <efried@us.ibm.com>
Closes-Bug: #1686488