Handle file delete races in image cache
In at least one case (see bug report) files have been deleted between
when we check that they exist and when we actually attempt the deletion.
Since we ultimately just want the file gone, we shouldn't care if we
get an error that the file does not exist and just move on.

Change-Id: Iaf71109fc6659650075ab4b4b48aec57819fb0b5
Closes-bug: #1229823