Add Discovery stub for Image Import
This change adds discovery stub for Image Import MVP returning
['glance-direct'] as import-methods if Image Import has been
enabled in config. The default behavior for Pike is disabled so
the GET call to /v2/info/import returns empty list as default.

The stub is required for the client work to support the Minimum
Viable Product testing.

This change is based on the work started by Cyril and Nikhil on
https://review.openstack.org/#/c/270980 dropping any components
that are depending of the parts of IIR not implemented yet.

Change-Id: I4bc4c1123fed7ea8a838ea3d06e9da966ae18f58
Co-Authored-By: Cyril Roelandt <cyril@redhat.com>
Co-Authored-By: Nikhil Komawar <nik.komawar@gmail.com>