Add 'api_image_import' type to task(s) schemas
The glanceclient relies on the schemas being accurate so it can
format responses.  Without this update, admins won't be able to
use the glanceclient generate a task list containing the
api_image_import task that the task engine uses to process the
new image import process being introduced in Pike.

Closes-bug: #1711018

Change-Id: I5bcc9f4cdc55635809e8a90be555a367348a58c2