Optimize the way to serach file 'glance-api-paste.ini'
With the original method _get_deployment_config_file() in config.py,
if the option config_file is specified in glance-api.conf, and run
command 'glance-api' under a directory, the the method load_paste_app()
will throw an IOError, but the IOError dose not been catched. The same
error will happen with'glance-registery'.

The reason for this IOError is the code "os.path.abspath(path)" in
_get_deployment_config_file()  will return a value
'{cur_dir}/glance-api-paste.ini', but the 'glance-api-paste.ini' does
not exist under {cur_dir}.Such as running the command under /opt, but
the 'glance-api-paste.ini' dose not exist under /opt.

This pacth modifies one line of code in method _get_paste_config_path()
for solving the IOError. At the same time, it provides one test case.

Change-Id: I970c1acb073700b15e153dd08c9ec14d20f0e83d
Closes-Bug: 1712226