Fix api_image_import tasks stuck in 'pending'
The immediate cause was that the api_image_import flow was not
listed in the entrypoints in setup.cfg, so stevedore couldn't find
it.  After the entrypoint was added, the task was going to failure
because it couldn't find the image_id.  So this patch has the
import controller add the image_id to the task_input dict.  Next
the flows were having trouble finding the image_id, so I just
passed it to them.  There's probably a more elegant way to have
taskflow handle this, so this patch could use a refactoring.  On
the plus side, the interoperable image import process does appear
to work with this patch.

NOTE: if you want to test this in devstack, use the stable/ocata
devstack but with the master branch of glance.  The tasks engine
doesn't appear to be running under the Pike devstack configuration.
(Use defect 1712463 if you have an idea what might be the problem.)

Change-Id: Ic68c17f4cb671eb664e2de331787b55fe9878a27
Closes-bug: #1712462
(cherry picked from commit 4366a7493c2d0020a70943f2ff78c663a99c81c3)