TrivialFix: Fix wrong test case
The 'md-property-create' API needs below required parameters to be
passed in the request body:
* type,
* title,
* name

If any of the parameter is absent in request body, glance-api returns
400 BadRequest with below error message:

InvalidObject(u"Provided object does not match schema 'property':
'<missing-property>' is a required property. Failed validating
'required' in schema: <API schema>
...
...
'required': ['type', 'title', 'name']

The below test case is written to validate length of name parameter
should not to be more than 255 characters:

glance.tests.unit.v2.test_metadef_resources.TestMetadefsControllers.
test_property_create_overlimit_name

This test is passing and raising HTTPBadRequest as expected but it's
not due to "name" parameter length exceeding max characters. It's
passing because this test is raising 400 error due to "type" parameter
not passed in the request body which is not the intention of this
particular unit test.

Fixed this issue by adding expected fake required attributes in
request body so that 'name' can be tested for maximum limit.

TrivialFix

Change-Id: I8dd18bfff881c4ebeae65fc7c183b750f2612df6