Scrubber refactor
The glance-scrubber utility is used by administrators
for the offline deletion of images when the Glance
option `delayed_delete` is enabled.  The refactoring in
this patch eliminates a dependency on the Glance
Registry v1 client. Further, after this change, the
glance-scrubber does not use the deprecated Glance
Registry at all. Instead, like the glance-manage tool,
it now visits the Glance database directly.

bp: scrubber-refactor
Change-Id: I26f570b85617200880543b7114730a1ac76d3fb1