Update http deploy docs to be a bit more explicit
There have been some lessons learned on how to best deploy glance from
using uwsgi in the gate. This commit updates the docs a bit to explain
some pitfalls to avoid.

This also adds a glossary at th end of the docs to explain common terms.
There are several very similar terms with related meanings used in the doc
so to clear up confusing the glossary should be helpful.

Change-Id: I255d2049a5d519909c88101833ddd259328fcee7