Clarify log message
The current message logged when showing image locations or direct_url
is enabled makes it sound like there's a problem with the storage
backend, whereas it's really just an explanation of why the locations
or direct_url are empty/missing in the image response.  This patch
rewrites the message to say that explicitly.

Change-Id: I9768fbba73aa290a7d159c01bd5ac0e9237f48ed
Closes-bug: #1728689