Remove setting of version/release from releasenotes
Release notes are version independent, so remove version/release values.
We've found that there are many projects now that require installation
in order to build release notes, and this is entirely due to the current
convention of pulling in the version information.

Release notes should not need installation in order to build, so this
unnecessary version setting needs to be removed.

Change-Id: If553414e9734b27b565c81f64845635edc0290eb
Needed-by: I56909152975f731a9d2c21b2825b972195e48ee8