Correct related section for enable_image_import
The documentation references the section DEFUALT
instead of DEFAULT. As this is most likely a
spelling error which will cause confusion, it is
better if we fix it.

This patch corrects the config option information
and includes an updated set of sample config files
generated from the current code including the fix.

Change-Id: If3c5e334aa1fa2ff5a28f52c00330d42cb9bcf9d