Prevent image become active without disk and container formats
If you run image-import api on any image which is in uploading
state and does not have container-format and/or disk-format set
goes into active state. The reason is during import call status
is set to importing from uploading and while checking container
format and disk formats are specified only 'queued' status was
taken into consideration.

Added 'uploading' status where validation is done to check whether
disk format and container format is added to the image.

Change-Id: I2865d81a01e5ab5d113bd809c273a2876213e7d4
Closes-Bug: #1733816